changed package.json to fire react server and api server
