+++ client/public/images/profilePictureSample.jpg
