+++ client/public/viewEventPage.html
