+++ client/src/components/ViewMyEventsPage/viewMyEventsHeader.js
