+++ client/src/components/mainpage/othercards.css
