+++ client/src/components/mainpage/statement.js
